This is the final code for OTA
Code checks the server version of code and compares. Fetches the code from server if latest code is available and the updates the ESP32 firmware with latest version and reboots, this code has two IR sensors to check the if passenger is sitting and Piezo sensor to measure the impact. if there is any collision detected then the corresponding air bag should be deployed for that we have added 3 LEDs to indicate.